Stokhos:  Remove old Fad design
